separate tests into a few different files
